hooks of ContentPanel put on own custom hook and returned from there
