+++ frontend/src/components/ContentPanel/index.tsx
@@ -1,5 +1,4 @@
-import { useState, FC, useEffect } from 'react';
-import { generateText } from '../../utils/generateContent';
+import { FC } from 'react';
 import {
     exportJson,
     downloadJson,
@@ -11,12 +10,9 @@ import { Surface } from '../Surface';
 import { ContentPanelHeader } from './ContentPanelHeader';
 import { ContentPanelPrompts } from './ContentPanelPrompts';
 import { ContentPanelActions } from './ContentPanelActions';
-import { InputSchema } from '../PromptIOBox';
 import classNames from 'classnames';
 import { Loader } from '../Loader';
-import { v4 as uuidv4 } from 'uuid';
-import { useAppDispatch } from '../../utils/hooks';
-import { updatePanel } from '../../reducers/panelReducer';
+import { usePanel } from './hooks';
 
 //Provide access to MasterCategory through a parent callback
 interface ContentPanelProps {
@@ -34,96 +30,18 @@ export const ContentPanel: FC<ContentPanelProps> = ({
     initialCategory,
     initialPrompts,
 }) => {
-    // Redux dispatch
-    const dispatch = useAppDispatch();
-
-    //Component state consists of category prompt, and N multiprompts (id, input, output)
-    const [category, setCategory] = useState<string>(initialCategory);
-    const [promptBoxes, setPromptBoxes] =
-        useState<PromptData[]>(initialPrompts);
-    const [loading, setLoading] = useState<boolean>(false);
-
-    //Callback to create new boxes in the panel
-    const addPromptBox = () => {
-        const newBox = { id: uuidv4(), input: '', output: '', locked: false };
-        setPromptBoxes((prev) => [...prev, newBox]);
-    };
-
-    // Generates all the IO boxes that are not locked
-    const generateAll = async () => {
-        // Start the loading spinner
-        setLoading(() => true);
-
-        // Map of <id, output> for content panels that are generated
-        const generated: Map<string, string> = new Map();
-
-        // Goes through the promptboxes and adds the map a new entry for ids which are not locked
-        for (const p of promptBoxes) {
-            // Generate if the prompt is not locked and input is valid
-            if (!p.locked && InputSchema.safeParse(p.input).success) {
-                const output: string = await generateText(
-                    p.id,
-                    p.input,
-                    category
-                );
-                generated.set(p.id, output);
-            }
-        }
-
-        // Sets all the promptboxes in a 1 setState call.
-        setPromptBoxes((prev) =>
-            prev.map((p) => {
-                if (generated.has(p.id)) {
-                    const mapOutput = generated.get(p.id);
-                    if (mapOutput) return { ...p, output: mapOutput };
-                }
-
-                return p;
-            })
-        );
-    };
-
-    /**
-     * A callback function for the setPromptBoxes which updates
-     * the redux store after the content generation
-     */
-    useEffect(() => {
-        if (loading) {
-            // Create the new panel object
-            const panel = {
-                id,
-                category,
-                prompts: promptBoxes,
-            };
-
-            // Update the redux store
-            dispatch(updatePanel(panel));
-            // Take out the loading spinner
-            setLoading(() => false);
-        }
-    }, [promptBoxes]);
-
-    // Generates single output
-    const generateOutput = async (p: PromptData) => {
-        setLoading(() => true);
-        setPromptOutput(p.id, await generateText(p.id, p.input, category));
-    };
-
-    //Callback to modify the input area of a PromptIOBox by id
-    const setPromptOutput = (id: string, output: string) => {
-        setPromptBoxes((prev) =>
-            prev.map((o) => (o.id === id ? { ...o, output: output } : o))
-        );
-    };
-
-    // Locks a prompt box, so that it is not generated with generate all
-    const lockPrompt = (id: string) => {
-        setPromptBoxes((prev) =>
-            prev.map((box) =>
-                box.id === id ? { ...box, locked: !box.locked } : box
-            )
-        );
-    };
+    const {
+        category,
+        promptBoxes,
+        loading,
+        setCategory,
+        setPromptBoxes,
+        generateAll,
+        generateOutput,
+        setPromptOutput,
+        lockPrompt,
+        addPromptBox,
+    } = usePanel(initialPrompts, initialCategory, id);
 
     //Callback to export the category, and all inputs / outputs in json
     const jsonExport = async () => {
